+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/apm/metrics/serverless/serverless_active_instances.spec.ts
@@ -6,15 +6,15 @@
  */
 
 import { APIReturnType } from '@kbn/apm-plugin/public/services/rest/create_call_apm_api';
+import type { ApmSynthtraceEsClient } from '@kbn/apm-synthtrace';
 import expect from '@kbn/expect';
 import { sumBy } from 'lodash';
-import { FtrProviderContext } from '../../../common/ftr_provider_context';
 import { config, expectedValues, generateData } from './generate_data';
+import { DeploymentAgnosticFtrProviderContext } from '../../../../../ftr_provider_context';
 
-export default function ApiTest({ getService }: FtrProviderContext) {
-  const registry = getService('registry');
-  const apmApiClient = getService('apmApiClient');
-  const apmSynthtraceEsClient = getService('apmSynthtraceEsClient');
+export default function ApiTest({ getService }: DeploymentAgnosticFtrProviderContext) {
+  const apmApiClient = getService('apmApi');
+  const synthtrace = getService('synthtrace');
 
   const start = new Date('2021-01-01T00:00:00.000Z').getTime();
   const end = new Date('2021-01-01T00:15:00.000Z').getTime() - 1;
@@ -36,8 +36,7 @@ export default function ApiTest({ getService }: FtrProviderContext) {
     });
   }
 
-  // FLAKY: https://github.com/elastic/kibana/issues/177639
-  registry.when('Serverless active instances', { config: 'basic', archives: [] }, () => {
+  describe('Serverless active instances', () => {
     const {
       memoryTotal,
       billedDurationMs,
@@ -47,8 +46,10 @@ export default function ApiTest({ getService }: FtrProviderContext) {
     } = config;
 
     const { expectedMemoryUsed } = expectedValues;
+    let apmSynthtraceEsClient: ApmSynthtraceEsClient;
 
     before(async () => {
+      apmSynthtraceEsClient = await synthtrace.createApmSynthtraceEsClient();
       await generateData({ start, end, apmSynthtraceEsClient });
     });
