+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/apm/metrics/serverless/serverless_functions_overview.spec.ts
@@ -7,13 +7,13 @@
 
 import { APIReturnType } from '@kbn/apm-plugin/public/services/rest/create_call_apm_api';
 import expect from '@kbn/expect';
-import { FtrProviderContext } from '../../../common/ftr_provider_context';
+import type { ApmSynthtraceEsClient } from '@kbn/apm-synthtrace';
 import { config, expectedValues, generateData } from './generate_data';
+import { DeploymentAgnosticFtrProviderContext } from '../../../../../ftr_provider_context';
 
-export default function ApiTest({ getService }: FtrProviderContext) {
-  const registry = getService('registry');
-  const apmApiClient = getService('apmApiClient');
-  const apmSynthtraceEsClient = getService('apmSynthtraceEsClient');
+export default function ApiTest({ getService }: DeploymentAgnosticFtrProviderContext) {
+  const apmApiClient = getService('apmApi');
+  const synthtrace = getService('synthtrace');
 
   const start = new Date('2021-01-01T00:00:00.000Z').getTime();
   const end = new Date('2021-01-01T00:15:00.000Z').getTime() - 1;
@@ -34,8 +34,7 @@ export default function ApiTest({ getService }: FtrProviderContext) {
     });
   }
 
-  // FLAKY: https://github.com/elastic/kibana/issues/177641
-  registry.when('Serverless functions overview', { config: 'basic', archives: [] }, () => {
+  describe('Serverless functions overview', () => {
     const {
       memoryTotal,
       billedDurationMs,
@@ -45,7 +44,10 @@ export default function ApiTest({ getService }: FtrProviderContext) {
     } = config;
     const { expectedMemoryUsed } = expectedValues;
 
+    let apmSynthtraceEsClient: ApmSynthtraceEsClient;
+
     before(async () => {
+      apmSynthtraceEsClient = await synthtrace.createApmSynthtraceEsClient();
       await generateData({ start, end, apmSynthtraceEsClient });
     });
